requirements.txt
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
crewai==0.130.0
crewai-tools==0.47.1
fastapi>=0.110.3
uvicorn
python-dotenv
python-multipart
pypdf
litellm
google-generativeai>=0.5.4
numpy>=1.26.4
openai>=1.30.5
pandas>=2.2.2
pillow>=10.3.0
pydantic>=2.0